// libs/data-access/src/lib/chats/interfaces/chats-ws-service.interface.ts
import { Observable } from 'rxjs';
import { ChatWsMessage } from './chat-ws-message.interface';

export interface ChatConnectionWsParams {
   url: string;
   token: string;
   handleMessage: (message: ChatWsMessage) => void;
}

export interface ChatWsService {
   connect: (
      params: ChatConnectionWsParams
   ) => void | Observable<ChatWsMessage>;
   sendMessage: (text: string, chatId: number) => void;
   disconnect: () => void;
}